Migrate from optparse to argparse
Optparse is deprecated since 2.7 and 3.2.

This kind of indirectly corrects a bug in jens-gc, as before this patch
and when not enabling --aggressive, None was being incorrectly passed to
git.gc().  With argparse, if the action of a parameter is store_true
then the default value is set to False.

https://docs.python.org/2.7/library/argparse.html#action