Return 200 if the hinted repository is not found
This is contrary to what we should be logically doing but Gitlab
considers hooks returning non-2xx as failed. It should be fine to say
"we don't know about this repository" via HTTP status codes, but c'est
la vie, I guess.

This plus [0] are a recipe for disaster. Anyway, even if [0] was not
the case disabling the hook for 10 minutes is still unacceptable for
Jens.

The workaround is to return 2xx, even in this case.

[0] https://gitlab.com/gitlab-org/gitlab/-/issues/387938